Remove ship mass (unnecessary feature, doesn't improve anything)
Initial idea (when the ship mass was first added) was for it to depend
on how much stuff was placed on the ship (blocks, buildings, etc.),
but this idea was abandoned, because there is no way to measure this
without heavy performance cost (full scan of the shipworld).

Currently ship mass is only increased by FTL Drives, which are mandatory
(player always has them), so increasing mass can't be avoided.

Player has methods to reduce the ship mass, but the mass only affects
fuel costs, so these methods can just reduce fuel cost instead.

In other words, ship mass is barely relevant to the player,
so we are removing this stat as unnecessary.